requirements.txt
# Web server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# YAML config loading
pyyaml==6.0.2

# Reddit API (narrative agent)
praw==7.8.1

# Twitter/X scraper (narrative agent — free, no API key)
twikit>=2.3.0

# Postgres driver (used when DATABASE_URL is set, ignored otherwise)
psycopg2-binary==2.9.10

# MCP server (AI agent discoverability for Claude/Cursor)
mcp>=1.26.0

# x402 payment gate (enabled when PAY_TO env var is set)
x402[fastapi,evm]==2.2.0

# CDP SDK — generates JWT auth tokens for Coinbase x402 facilitator
cdp-sdk>=1.39.0

# Phase D: ML pipeline (meta-learner, meta-labeler, calibration)
lightgbm>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
